geolocator/model.py
from pathlib import Path
from typing import Dict, Any, Tuple, List, Optional
import os
import random

import numpy as np
import torch
from PIL import Image


class GeoPredictor:
    """Predict image location using a simplified model."""
    
    def __init__(self, model_name: str = ""):
        """Initialize the geolocation predictor.
        
        Args:
            model_name: Unused for now, kept for API compatibility
        """
        self.is_loaded = True
        
        # Predefined regions (we'll expand this or use a more granular approach)
        self.regions = {
            0: {"name": "North America", "center": (40.7128, -74.0060)},
            1: {"name": "South America", "center": (-15.7801, -47.9292)},
            2: {"name": "Europe", "center": (48.8566, 2.3522)},
            3: {"name": "Africa", "center": (-1.2921, 36.8219)},
            4: {"name": "Asia", "center": (35.6762, 139.6503)},
            5: {"name": "Australia", "center": (-33.8688, 151.2093)},
        }
    
    def predict(self, image_array: np.ndarray) -> Tuple[Tuple[float, float], Dict[str, Any]]:
        """Predict location based on image features.
        
        Args:
            image_array: Preprocessed image as numpy array
            
        Returns:
            Tuple of (latitude, longitude) and additional prediction metadata
        """
        # For this simplified version, we'll analyze basic image features
        # without using a complex model, and make a simple prediction
        
        # Calculate average brightness and color distribution
        brightness = np.mean(image_array)
        
        # Get color distribution (simplified)
        # Higher values in red channel might suggest desert or urban areas
        # Higher values in green channel might suggest forests or grasslands
        # Higher values in blue channel might suggest water or sky
        red_avg = np.mean(image_array[:, :, 0])
        green_avg = np.mean(image_array[:, :, 1])
        blue_avg = np.mean(image_array[:, :, 2])
        
        # Simple logic for region selection
        if green_avg > red_avg and green_avg > blue_avg:
            # Likely forests or grasslands
            region_idx = 0  # North America
        elif red_avg > green_avg and red_avg > blue_avg:
            # Likely desert or urban
            region_idx = 4  # Asia
        elif blue_avg > red_avg and blue_avg > green_avg:
            # Likely water or beaches
            region_idx = 5  # Australia
        else:
            # Otherwise, choose a random region
            region_idx = random.randint(0, len(self.regions) - 1)
        
        region = self.regions[region_idx]
        
        # Calculate a fake confidence based on the strength of the color differences
        max_diff = max(abs(red_avg - green_avg), abs(red_avg - blue_avg), abs(green_avg - blue_avg))
        confidence = min(0.7, 0.3 + max_diff * 2)  # Reasonable confidence range
        
        # Return predicted coordinates and metadata
        metadata = {
            "region": region["name"],
            "confidence": confidence,
            "predicted_class": int(region_idx)
        }
        
        return region["center"], metadata
        
    def get_region_info(self, lat: float, lon: float) -> Dict[str, Any]:
        """Get information about the region containing the coordinates.
        
        Args:
            lat: Latitude
            lon: Longitude
            
        Returns:
            Dictionary with region information
        """
        # Simple implementation - find closest region center
        min_dist = float('inf')
        closest_region = None
        
        for region_id, region_data in self.regions.items():
            region_lat, region_lon = region_data["center"]
            
            # Simple Euclidean distance (not accurate for geographic coords but sufficient for demo)
            dist = ((lat - region_lat) ** 2 + (lon - region_lon) ** 2) ** 0.5
            
            if dist < min_dist:
                min_dist = dist
                closest_region = region_data
                
        return closest_region