pyproject.toml
[project]
name = "geolocator"
version = "0.1.0"
description = "Image-based geolocation predictor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow",
    "exifread",
    "numpy",
    "requests",
    "typer",
    "rich",
    "pytest"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov"
]
